+++ packages/webos/CHANGELOG.md
@@ -7,7 +7,7 @@ The following is a curated list of changes in the Enact webos module, newest cha
 ### Added
 
 - `onTimeout` and `timeout` props to `LS2Request`
-- default `onTimeout` handler to `LS2Request`
+- default `onFailure` and `onTimeout` handlers to `LS2Request`
 
 ## [2.0.0-alpha.6] - 2018-03-22